Specular lighting added. Changed relative paths so that they work without typing the full absolute path. Removed allCharsFromFile leak in ATScene.cpp
TODO: fix memory leak with ATTextureMap pixelMap, fix blending by culling back faces

+++ ATShader/ATScene.h
@@ -36,7 +36,7 @@ public:
     
     void splitString(string line, string retVals[], unsigned int count);
     
-    vector<ATVector4D> createVectors(char allCharsFromFile[], unsigned int count, float* totalX, float* totalY, int* numVals);
+    vector<ATVector4D> createVectors(char allCharsFromFile[], unsigned long count, float* totalX, float* totalY, int* numVals);
     
     bool isLineWhitespace(string line);
